# rpc/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Tendermint RPC
  contact:
    name: Tendermint RPC
    url: https://github.com/providenetwork/tendermint/issues/new/choose
  description: |
    Tendermint supports the following RPC protocols:

    * URI over HTTP
    * JSONRPC over HTTP
    * JSONRPC over websockets

    ## Configuration

    RPC can be configured by tuning parameters under `[rpc]` table in the
    `$TMHOME/config/config.toml` file or by using the `--rpc.X` command-line
    flags.

    Default rpc listen address is `tcp://0.0.0.0:26657`.
    To set another address, set the `laddr` config parameter to desired value.
    CORS (Cross-Origin Resource Sharing) can be enabled by setting
    `cors_allowed_origins`, `cors_allowed_methods`, `cors_allowed_headers`
    config parameters.

    ## Arguments

    Arguments which expect strings or byte arrays may be passed as quoted
    strings, like `"abc"` or as `0x`-prefixed strings, like `0x616263`.

    ## URI/HTTP

    A REST like interface.

        curl localhost:26657/block?height=5

    ## JSONRPC/HTTP

    JSONRPC requests can be POST'd to the root RPC endpoint via HTTP.

        curl --header "Content-Type: application/json" --request POST --data '{"method": "block", "params": ["5"], "id": 1}' localhost:26657

    ## JSONRPC/websockets

    JSONRPC requests can be also made via websocket.
    The websocket endpoint is at `/websocket`, e.g. `localhost:26657/websocket`.
...
    only available via websockets.

...

...
  version: "Master"
...
servers:
...
paths:
...
  /block_results:
    get:
      summary: Get block results at a specified height
      operationId: block_results
      parameters:
        - in: query
          name: height
          description: height to return. If no height is provided, it will fetch informations regarding the latest block.
          schema:
            type: integer
            default: 0
            example: 1
      tags:
        - Info
      description: |
        Get block_results.
      responses:
        "200":
          description: Block results.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockSearchResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /validators:
    get:
      summary: Get validator set at a specified height
      operationId: validators
      parameters:
        - in: query
          name: height
          description: height to return. If no height is provided, it will fetch validator set which corresponds to the latest block.
          schema:
            type: integer
            default: 0
            example: 1
        - in: query
          name: page
          description: "Page number (1-based)"
          required: false
          schema:
            type: integer
            default: 1
            example: 1
        - in: query
          name: per_page
          description: "Number of entries per page (max: 100)"
          required: false
          schema:
            type: integer
            example: 30
            default: 30
      tags:
        - Info
      description: |
        Get Validators. Validators are sorted first by voting power (descending), then by address (ascending).
      responses:
        "200":
          description: Commit results.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorsResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /consensus_params:
    get:
      summary: Get consensus parameters
      operationId: consensus_params
      parameters:
        - in: query
          name: height
          description: height to return. If no height is provided, it will fetch commit informations regarding the latest block.
          schema:
            type: integer
            default: 0
            example: 1
      tags:
        - Info
      description: |
        Get consensus parameters.
      responses:
        "200":
          description: consensus parameters results.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsensusParamsResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    JSONRPC:
      type: object
      properties:
        id:
          type: integer
          example: 0
        jsonrpc:
          type: string
          example: "2.0"
...
    ErrorResponse:
      description: Error Response
      allOf:
        - $ref: "#/components/schemas/JSONRPC"
        - type: object
          properties:
            error:
              type: string
              example: "Description of failure"
...
    PubKey:
      type: object
      properties:
        type:
          type: string
          example: "tendermint/PubKeyEd25519"
        value:
          type: string
          example: "A6DoBUypNtUAyEHWtQ9bFjfNg8Bo9CrnkUGl6k6OHN4="
...
    Commit:
      required:
        - "type"
        - "height"
        - "round"
        - "block_id"
        - "timestamp"
        - "validator_address"
        - "validator_index"
        - "signature"
      properties:
        type:
          type: integer
          example: 2
        height:
          type: string
          example: "1262085"
        round:
          type: integer
          example: 0
        block_id:
          $ref: "#/components/schemas/BlockID"
        timestamp:
          type: string
          example: "2019-08-01T11:39:38.867269833Z"
        validator_address:
          type: string
          example: "000001E443FD237E4B616E2FA69DF4EE3D49A94F"
        validator_index:
          type: integer
          example: 0
        signature:
          type: string
          example: "DBchvucTzAUEJnGYpNvMdqLhBAHG4Px8BsOBB3J3mAFCLGeuG7uJqy+nVngKzZdPhPi8RhmE/xcw/M9DOJjEDg=="
    Block:
      type: object
      properties:
        header:
          $ref: "#/components/schemas/BlockHeader"
        data:
          type: array
          items:
            type: string
            example: "yQHwYl3uCkKoo2GaChRnd+THLQ2RM87nEZrE19910Z28ABIUWW/t8AtIMwcyU0sT32RcMDI9GF0aEAoFdWF0b20SBzEwMDAwMDASEwoNCgV1YXRvbRIEMzEwMRCd8gEaagom61rphyEDoJPxlcjRoNDtZ9xMdvs+lRzFaHe2dl2P5R2yVCWrsHISQKkqX5H1zXAIJuC57yw0Yb03Fwy75VRip0ZBtLiYsUqkOsPUoQZAhDNP+6LY+RUwz/nVzedkF0S29NZ32QXdGv0="
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/Evidence"
        last_commit:
          type: object
          properties:
            height:
              type: integer
            round:
              type: integer
            block_id:
              $ref: "#/components/schemas/BlockID"
            signatures:
              type: array
              items:
                $ref: "#/components/schemas/Commit"
    Evidence:
      type: object
      properties:
        type:
          type: string
        height:
          type: integer
        time:
          type: integer
        total_voting_power:
          type: integer
        validator:
          $ref: "#/components/schemas/Validator"
    BlockComplete:
      type: object
      properties:
        block_id:
          $ref: "#/components/schemas/BlockID"
        block:
          $ref: "#/components/schemas/Block"
...
    ValidatorsResponse:
      type: object
      required:
        - "jsonrpc"
        - "id"
        - "result"
      properties:
        jsonrpc:
          type: string
          example: "2.0"
        id:
          type: integer
          example: 0
        result:
          required:
            - "block_height"
            - "validators"
          properties:
            block_height:
              type: string
              example: "55"
            validators:
              type: array
              items:
                $ref: "#/components/schemas/ValidatorPriority"
            count:
              type: string
              example: "1"
            total:
              type: string
              example: "25"
          type: object
...
    ConsensusParamsResponse:
      type: object
      required:
        - "jsonrpc"
        - "id"
        - "result"
      properties:
        jsonrpc:
          type: string
          example: "2.0"
        id:
          type: integer
          example: 0
        result:
          type: object
          required:
            - "block_height"
            - "consensus_params"
          properties:
            block_height:
              type: string
              example: "1"
            consensus_params:
              $ref: "#/components/schemas/ConsensusParams"
...
    BlockSearchResponse:
      type: object
      required:
        - "jsonrpc"
        - "id"
        - "result"
      properties:
        jsonrpc:
          type: string
          example: "2.0"
        id:
          type: integer
          example: 0
        result:
          required:
            - "blocks"
            - "total_count"
          properties:
            blocks:
              type: array
              items:
                $ref: "#/components/schemas/BlockComplete"
            total_count:
              type: integer
              example: 2
          type: object

    ###### Reuseable types ######

    # Validator type with proposer prioirty
    ValidatorPriority:
      type: object
      properties:
        address:
          type: string
          example: "000001E443FD237E4B616E2FA69DF4EE3D49A94F"
        pub_key:
          required:
            - "type"
            - "value"
          properties:
            type:
              type: string
              example: "tendermint/PubKeyEd25519"
            value:
              type: string
              example: "9tK9IT+FPdf2qm+5c2qaxi10sWP+3erWTKgftn2PaQM="
          type: object
        voting_power:
          type: string
          example: "239727"
        proposer_priority:
          type: string
          example: "-11896414"

    # Stripped down validator
    Validator:
      type: object
      properties:
        pub_key:
          $ref: "#/components/schemas/PubKey"
        voting_power:
          type: integer
        address:
          type: string

    # Consensus Params
    ConsensusParams:
      type: object
      nullable: true
      required:
        - "block"
        - "evidence"
        - "validator"
      properties:
        block:
          type: object
          required:
            - "max_bytes"
            - "max_gas"
            - "time_iota_ms"
          properties:
            max_bytes:
              type: string
              example: "22020096"
            max_gas:
              type: string
              example: "1000"
            time_iota_ms:
              type: string
              example: "1000"
        evidence:
          type: object
          required:
            - "max_age"
          properties:
            max_age:
              type: string
              example: "100000"
        validator:
          type: object
          required:
            - "pub_key_types"
          properties:
            pub_key_types:
              type: array
              items:
                type: string
              example:
                - "ed25519"

    # Events in tendermint
...
    BlockHeader:
      required:
        - "version"
        - "chain_id"
        - "height"
        - "time"
        - "last_block_id"
        - "last_commit_hash"
        - "data_hash"
        - "validators_hash"
        - "next_validators_hash"
        - "consensus_hash"
        - "app_hash"
        - "last_results_hash"
        - "evidence_hash"
        - "proposer_address"
      properties:
        version:
          required:
            - "block"
            - "app"
          properties:
            block:
              type: string
              example: "10"
            app:
              type: string
              example: "0"
          type: object
        chain_id:
          type: string
          example: "cosmoshub-2"
        height:
          type: string
          example: "12"
        time:
          type: string
          example: "2019-04-22T17:01:51.701356223Z"
        last_block_id:
          $ref: "#/components/schemas/BlockID"
        last_commit_hash:
          type: string
          example: "21B9BC845AD2CB2C4193CDD17BFC506F1EBE5A7402E84AD96E64171287A34812"
        data_hash:
          type: string
          example: "970886F99E77ED0D60DA8FCE0447C2676E59F2F77302B0C4AA10E1D02F18EF73"
        validators_hash:
          type: string
          example: "D658BFD100CA8025CFD3BECFE86194322731D387286FBD26E059115FD5F2BCA0"
        next_validators_hash:
          type: string
          example: "D658BFD100CA8025CFD3BECFE86194322731D387286FBD26E059115FD5F2BCA0"
        consensus_hash:
          type: string
          example: "0F2908883A105C793B74495EB7D6DF2EEA479ED7FC9349206A65CB0F9987A0B8"
        app_hash:
          type: string
          example: "223BF64D4A01074DC523A80E76B9BBC786C791FB0A1893AC5B14866356FCFD6C"
        last_results_hash:
          type: string
          example: ""
        evidence_hash:
          type: string
          example: ""
        proposer_address:
          type: string
          example: "D540AB022088612AC74B287D076DBFBC4A377A2E"
      type: object
    BlockID:
      required:
        - "hash"
        - "parts"
      properties:
        hash:
          type: string
          example: "112BC173FD838FB68EB43476816CD7B4C6661B6884A9E357B417EE957E1CF8F7"
        parts:
          required:
            - "total"
            - "hash"
          properties:
            total:
              type: integer
              example: 1
            hash:
              type: string
              example: "38D4B26B5B725C4F13571EFE022C030390E4C33C8CF6F88EDD142EA769642DBD"
          type: object